Move NetworkPolicy handlers to their own package
