+++ pkg/controller/networkpolicy/networkpolicy_test.go
@@ -1,4 +1,4 @@
-package appdefinition
+package networkpolicy
 
 import (
 	"testing"
